+++ packages/wallet-connectors/src/WalletConnect.ts
@@ -188,19 +188,12 @@ export class WalletConnectConnection implements WalletConnection {
     }
 
     /**
-     * Non-async variant of {@link getConnectedAccount}.
-     * The async version is a simple wrapper around this one and
-     * only exists to satisfy the {@link WalletConnection} interface.
-     * So prefer this method when interacting with this concrete type.
+     * @return The account that the wallet currently associates with this connection.
      */
-    getConnectedAccount_() {
+    getConnectedAccount() {
         // We're only expecting a single account to be connected.
         const fullAddress = this.session.namespaces[WALLET_CONNECT_SESSION_NAMESPACE].accounts[0];
         return fullAddress.substring(fullAddress.lastIndexOf(':') + 1);
-    }
-
-    async getConnectedAccount() {
-        return this.getConnectedAccount_();
     }
 
     getJsonRpcClient() {
@@ -318,11 +311,7 @@ export class WalletConnectConnector implements WalletConnector {
             const { namespaces } = params;
             // Overwrite session.
             connection.session = { ...connection.session, namespaces };
-            // TODO Only fire event if the account actually changed.
-            connection
-                .getConnectedAccount()
-                .then((a) => delegate.onAccountChanged(connection, a))
-                .catch(console.error);
+            delegate.onAccountChanged(connection, connection.getConnectedAccount());
         });
         client.on('session_delete', ({ topic }) => {
             // Session was deleted: Reset the dApp state, clean up user session, etc.
@@ -364,7 +353,7 @@ export class WalletConnectConnector implements WalletConnector {
         const rpcClient = new JsonRpcClient(new HttpProvider(this.network.jsonRpcUrl));
         const connection = new WalletConnectConnection(this, rpcClient, chainId, session);
         this.connections.set(session.topic, connection);
-        this.delegate.onConnected(connection, connection.getConnectedAccount_());
+        this.delegate.onConnected(connection, connection.getConnectedAccount());
         return connection;
     }